requirements.txt
black==21.9b0
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
Flask==2.0.2
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.2
MarkupSafe==2.0.1
mypy-extensions==0.4.3
numpy==1.21.2
pathspec==0.9.0
platformdirs==2.4.0
PyAudio==0.2.11
pyserial==3.5
regex==2021.10.8
requests==2.26.0
scipy==1.7.1
six==1.16.0
spotipy==2.19.0
tomli==1.2.1
typing-extensions==3.10.0.2
urllib3==1.26.7
Werkzeug==2.0.2
